Move the BridgeTree type to the root module.
